% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_bananagram}
\alias{dat_bananagram}
\title{Counts of letters from Banangrams.}
\format{A data frame with three variables: \code{letter}, \code{bag_1}, and \code{bag_2}

\describe{
\item{\code{letter}}{letters from A-Z}
\item{\code{bag_1}}{Number of each letter in bag 1}
\item{\code{bag_2}}{Number of each letter in bag 2}
}}
\usage{
dat_bananagram
}
\description{
Counts of the letters from two randomly sorted bags of bananagrams.
}
\keyword{datasets}
